requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.4
python-dotenv==1.0.0
requests==2.31.0
PyJWT==2.8.0
python-engineio==4.5.1
python-socketio==5.8.0
eventlet==0.33.3
